refactor(ooda_actions): split tests_goal_session_inline (511 → 319 + 200)
Refs #1266
